split feed list ui into seperate files and directories for modals
