+++ Modules/feed/Views/feed_list.css
@@ -1,0 +1,191 @@
+
+body{padding:0!important}
+
+#footer {
+    margin-left: 0px;
+    margin-right: 0px;
+}
+
+.controls { margin-bottom:10px; }
+
+#feeds-to-delete { font-style:italic; }
+
+#deleteFeedModalSelectedItems{
+    position:absolute;
+    overflow:hidden;
+    text-align:left;
+    background: #f5f5f5;
+}
+#deleteFeedModalSelectedItems h5{ margin:0 }
+#deleteFeedModalSelectedItems ol{
+    max-width:80%;
+    position:absolute;
+}
+
+/* CSS Grid Feed List Styles */
+.feed-list-grid {
+    display: grid;
+    grid-template-columns: 40px minmax(max-content, 2fr) 40px max-content 80px minmax(max-content, 1fr) 80px 100px;
+    width: 100%;
+    box-sizing: border-box;
+    contain: layout style; /* CSS containment for better performance */
+    gap: 0; /* Explicit gap control */
+}
+
+.grid-row {
+    display: grid;
+    grid-column: 1 / -1;
+    grid-template-columns: subgrid;
+    align-items: center;
+    min-height: 41px;
+    cursor: default;
+}
+
+/* Responsive behavior */
+@media (max-width: 992px) {
+    .feed-list-grid {
+        grid-template-columns: 40px max-content 40px max-content minmax(80px, 1fr) 80px 100px;
+    }
+    
+    /* Start by hiding the process list column */
+    .grid-row > .grid-cell:nth-child(6) {
+        display: none;
+    }
+}
+
+@media (max-width: 768px) {
+    .feed-list-grid {
+        grid-template-columns: 40px 1fr 80px 80px;
+    }
+    /* Hide public, engine, and size columns */
+    .grid-row > .grid-cell:nth-child(3),  /* Public column */
+    .grid-row > .grid-cell:nth-child(4),  /* Engine column */
+    .grid-row > .grid-cell:nth-child(5) { /* Size column */
+        display: none;
+    }
+}
+
+@media (max-width: 480px) {
+    .feed-list-grid {
+        grid-template-columns: 30px 1fr 80px;
+    }
+    /* Additionally hide value column on very small screens */
+    .grid-row > .grid-cell:nth-child(8) {
+        display: none;
+    }
+}
+
+.feed-header {
+    background: #ddd;
+    font-weight: bold;
+    border-bottom: 1px solid #ccc;
+}
+
+.grid-cell {
+    padding: 4px 8px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    min-width: 0;
+    /*Not sure if this adds anything*/
+    word-break: break-word;
+    hyphens: auto;
+}
+
+.text-center {
+    text-align: center;
+}
+
+.text-left {
+    text-align: left;
+}
+
+/* Node header styles for grid */
+.node-header {
+    background: #ddd;
+    cursor: pointer;
+    transition: background-color 0.15s ease-in-out;
+    border-right: 4px solid var(--status-color);
+}
+
+.node-header:hover {
+    background-color: #e3e3e3;
+}
+
+.node-header h5 {
+    margin: 0;
+    font-weight: bold;
+    font-size: 14px;
+    grid-column: 2;
+}
+
+/* Feed row styles for grid */
+.node-feed {
+    border-bottom: 1px solid #fff;
+    background: #f0f0f0;
+    position: relative;
+    transition: background-color 0.1s ease;
+    width: 100%;
+    cursor: pointer;
+    border-right: 4px solid var(--status-color);
+}
+
+.node-feed:hover {
+    background-color: #f5f5f5;
+}
+
+.node-feed:last-of-type {
+    border-bottom-width: 0;
+}
+
+/* Feed status indicator on right side */
+.node-feed:after {
+    content: '';
+    width: 0px;
+    background: var(--bg-menu-top);
+    height: 100%;
+    display: block;
+    position: absolute;
+    left: 0;
+    top: 0;
+    transition: width .2s ease-out;
+}
+
+.node-feed:hover:after {
+    width: 2px;
+}
+
+/* Selected feed styling */
+.node-feed.selected {
+    background-color: rgba(60, 135, 170, 0.2) !important; /* Semi-transparent blue background */
+}
+
+.node-feed.selected:after {
+    width: 4px !important;
+    background: rgb(60, 135, 170) !important; /* Blue left border */
+}
+
+.node-feed.selected:hover {
+    background-color: rgba(60, 135, 170, 0.15) !important; /* Slightly more opaque on hover */
+}
+
+/* Arrow animation for Vue */
+.arrow-icon {
+    transition: transform 0.3s ease;
+    opacity: 0.333;
+}
+
+/* Collapsible content for Vue */
+.vue-collapsible-content {
+    grid-column: 1 / -1;
+    display: grid;
+    grid-template-columns: subgrid;
+    max-height: 0;
+    overflow: hidden;
+    transition: max-height 0.2s ease-in-out;
+    will-change: max-height;
+}
+
+.vue-collapsible-content.is-expanded {
+    max-height: 2000px; /* Adjust as needed, should be larger than any possible content height */
+}
