+++ Modules/feed/Views/modals/downsample/downsample_modal.js
