+++ Modules/feed/Views/modals/new/new_modal.js
@@ -1,0 +1,60 @@
+
+// ---------------------------------------------------------------------------------------------
+// Create new feed dialog
+// ---------------------------------------------------------------------------------------------
+
+for (var e in engines_hidden) {
+    $('#newfeed-engine option[value='+engines_hidden[e]+']').hide();
+}
+
+$("#newfeed-save").click(function (){
+    var name = $('#newfeed-name').val();
+    var tag = $('#newfeed-tag').val();
+    var engine = $('#newfeed-engine').val();
+    
+    var options = {};
+    if (engine==5) {
+        options.interval = $('#newfeed-interval').val();
+    }
+    
+    var result = feed.create(tag,name,engine,options);
+    feedid = result.feedid;
+
+    if (!result.success || feedid<1) {
+        alert('ERROR: Feed could not be created. '+result.message);
+        return false;
+    } else {
+        update_feed_list(); 
+        $('#newFeedNameModal').modal('hide');
+    }
+});
+
+$('#newfeed-engine').change(function(){
+    var engine = $(this).val();
+    if (engine==5) {
+        $('#newfeed-interval').show();
+    } else {
+        $('#newfeed-interval').hide();
+    }
+});
+
+$(".feed-process").click(function() {
+    // There should only ever be one feed that is selected here:
+    var feedid = 0; for (var z in selected_feeds) { if (selected_feeds[z]) feedid = z; }
+    var contextid = feedid;
+    var contextname = "";
+    if (feeds[feedid].name != "") contextname = feeds[feedid].tag + " : " + feeds[feedid].name;
+    else contextname = feeds[feedid].tag + " : " + feeds[feedid].id;    
+    process_vue.load(1, contextid, feeds[feedid].processList, contextname, null, null); // load configs
+});
+
+function save_processlist(feed_id, process_list) {
+    var result = feed.set_process(feed_id, process_list);
+    if (!result.success) {
+        alert('ERROR: Could not save processlist. '+result.message); 
+        return false;
+    } else {
+        update_feed_list();
+        return true;
+    }
+}
